Extract form methods into utility file
